Chunk 4 done?
Hi Andrew, 
I have added the code I did for chunk 4

I did something similar to what we talked about yesterday, with a slightly different approach: I joined "votings" with "votings" to get A-->B and B-->C and then compared with the original "votings" to check whether A-->C or not (that's what the AtoC columns are about).

After filtering these to get the closed directed triangles I removed a few useless columns and filtered by timestamp and also removed votes with zeros

Then I created new columns to include which statement each column refers to ("pp" = "the friend of my friend", "pn"= the friend of my enemy" and so on) and also the result of that operation (pp is +1 *+1 = +1, etc). Then in the last column I compare vote AC with statement AC to see if we really see that behaviour.

And finally I group everything by statement to check how accurate each statement is.

Makes sense?