Chunks 4, 5 and 6 added
I think all chunks are correct. Still missing the visualization part in chunk 4, although I get the confirmation on how many occurences of each are and whether they hold true most of the time or not. It seems they only do for "the friend of my friend is my friend"

I used the knowledge from chunk 4 to create the graph in chunk 5 with the properties of the edges already there. I also did the sorting of "votings2" by timestamp to assign the train and test datasets.

Chunk 6 is exactly as chunk 2 only changing "gtrain" by "gtrain2" and "gtest" by "gtest2". Edges have all the properties now

Hopefully it makes sense!

BR
Carlos